backend/requirements.txt
# Core
fastapi[all]>=0.109.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.6.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
asyncpg>=0.29.0

# AI & API
groq>=0.4.2
httpx>=0.26.0
python-dotenv>=1.0.1


# Utils
structlog>=24.1.0  # Structured logging
orjson>=3.9.12    # Fast JSON
rich>=13.7.0      # Beautiful logs

# Extraction
PyMuPDF>=1.23.0
python-docx>=1.1.0
pillow>=10.2.0
python-multipart>=0.0.7
